sync: Move barrier definitions
