+++ BUILD.gn
@@ -345,6 +345,8 @@ vvl_sources = [
   "layers/sync/sync_access_flags.h",
   "layers/sync/sync_access_state.cpp",
   "layers/sync/sync_access_state.h",
+  "layers/sync/sync_barrier.cpp",
+  "layers/sync/sync_barrier.h",
   "layers/sync/sync_commandbuffer.cpp",
   "layers/sync/sync_commandbuffer.h",
   "layers/sync/sync_common.cpp",
