+++ layers/CMakeLists.txt
@@ -433,6 +433,8 @@ target_sources(vvl PRIVATE
     sync/sync_access_flags.h
     sync/sync_access_state.cpp
     sync/sync_access_state.h
+    sync/sync_barrier.cpp
+    sync/sync_barrier.h
     sync/sync_commandbuffer.cpp
     sync/sync_commandbuffer.h
     sync/sync_common.cpp
