+++ layers/sync/sync_access_state.h
@@ -17,6 +17,7 @@
 
 #pragma once
 #include "sync/sync_common.h"
+#include "sync/sync_barrier.h"
 #include "containers/span.h"
 #include <cstring>  // memcpy
 
@@ -24,6 +25,7 @@ class ResourceAccessState;
 struct WriteState;
 struct ReadState;
 struct ResourceFirstAccess;
+struct SemaphoreScope;
 
 namespace syncval_stats {
 struct AccessContextStats;
@@ -143,58 +145,6 @@ class HazardResult {
     std::optional<HazardState> state_;
 };
 
-struct SyncExecScope {
-    VkPipelineStageFlags2 mask_param = 0;  // the xxxStageMask parameter passed by the caller
-    VkPipelineStageFlags2 exec_scope = 0;  // all earlier or later stages that would be affected by a barrier using this scope.
-    SyncAccessFlags valid_accesses;        // all valid accesses that can be used with this scope.
-
-    static SyncExecScope MakeSrc(VkQueueFlags queue_flags, VkPipelineStageFlags2 src_stage_mask,
-                                 const VkPipelineStageFlags2 disabled_feature_mask = 0);
-    static SyncExecScope MakeDst(VkQueueFlags queue_flags, VkPipelineStageFlags2 src_stage_mask);
-};
-
-struct SemaphoreScope : SyncExecScope {
-    SemaphoreScope(QueueId qid, const SyncExecScope &exec_scope) : SyncExecScope(exec_scope), queue(qid) {}
-    SemaphoreScope() = default;
-    QueueId queue;
-};
-
-struct SyncBarrier {
-    struct AllAccess {};
-    SyncExecScope src_exec_scope;
-    SyncAccessFlags src_access_scope;
-    SyncExecScope dst_exec_scope;
-    SyncAccessFlags dst_access_scope;
-
-    SyncBarrier() = default;
-    SyncBarrier(const SyncExecScope &src_exec, const SyncExecScope &dst_exec);
-    SyncBarrier(const SyncExecScope &src_exec, const SyncExecScope &dst_exec, const AllAccess &);
-    SyncBarrier(const SyncExecScope &src_exec, VkAccessFlags2 src_access_mask, const SyncExecScope &dst_exec,
-                VkAccessFlags2 dst_access_mask);
-    SyncBarrier(VkQueueFlags queue_flags, const VkSubpassDependency2 &barrier);
-    SyncBarrier(const std::vector<SyncBarrier> &barriers);
-};
-
-// Defines the source scope of the barrier.
-// ReadState and WriteState have InBarrierSourceScope() that checks if corresponding access is in the barrier source scope.
-struct BarrierScope {
-    VkPipelineStageFlagBits2 src_exec_scope;
-    SyncAccessFlags src_access_scope;
-
-    // Scope queue is used to include accesses only from the specific queue.
-    // The check against queue scope is unified for all cases. During record time the scope queue
-    // has default value (Invalid). This matches how the queue member of read/write accesses is
-    // initialized during recording, so (access_queue == scope_queue) evaluates to true during record time.
-    QueueId scope_queue = kQueueIdInvalid;
-
-    // The tag is needed for the event scope logic. The scope tag is defined by the "set event" command.
-    // The check against scope tag is unified for all cases. For non event code the scope tag is uint64-max
-    // value, so (access_tag < scope_tag) evaluates to true for non event code.
-    ResourceUsageTag scope_tag = kInvalidTag;
-
-    BarrierScope(const SyncBarrier &barrier, QueueId scope_queue = kQueueIdInvalid, ResourceUsageTag scope_tag = kInvalidTag);
-};
-
 struct ResourceFirstAccess {
     const SyncAccessInfo *usage_info;
     ResourceUsageTag tag;
@@ -251,19 +201,9 @@ struct ReadState {
         return (stage == rhs.stage) && (access_index == rhs.access_index) && (barriers == rhs.barriers) &&
                (sync_stages == rhs.sync_stages) && (tag == rhs.tag) && (queue == rhs.queue);
     }
-    bool IsReadBarrierHazard(VkPipelineStageFlags2 src_exec_scope) const {
-        // If the read stage is not in the src sync scope
-        // *AND* not execution chained with an existing sync barrier (that's the or)
-        // then the barrier access is unsafe (R/W after R)
-        return (src_exec_scope & (stage | barriers)) == 0;
-    }
+
     bool IsReadBarrierHazard(QueueId barrier_queue, VkPipelineStageFlags2 src_exec_scope,
                              const SyncAccessFlags &src_access_scope) const {
-        // If the read stage is not in the src sync scope
-        // *AND* not execution chained with an existing sync barrier (that's the or)
-        // then the barrier access is unsafe (R/W after R)
-        VkPipelineStageFlags2 queue_ordered_stage = (queue == barrier_queue) ? stage : VK_PIPELINE_STAGE_2_NONE;
-
         // Current implementation relies on TOP_OF_PIPE constant due to the fact that it's non-zero value
         // and AND-ing with it can create execution dependency when it's necessary. When NONE constant is
         // used, which equals to zero, then AND-ing with it always results in 0 which means "no barrier",
@@ -277,6 +217,10 @@ struct ReadState {
             src_exec_scope = VK_PIPELINE_STAGE_2_TOP_OF_PIPE_BIT;
         }
 
+        // If the read stage is not in the src sync scope
+        // *AND* not execution chained with an existing sync barrier (that's the or)
+        // then the barrier access is unsafe (R/W after R)
+        VkPipelineStageFlags2 queue_ordered_stage = (queue == barrier_queue) ? stage : VK_PIPELINE_STAGE_2_NONE;
         return (src_exec_scope & (queue_ordered_stage | barriers)) == 0;
     }
     bool ReadOrDependencyChainInSourceScope(QueueId queue, VkPipelineStageFlags2 src_exec_scope) const;
@@ -421,7 +365,7 @@ class ResourceAccessState {
     void ApplyPendingWriteBarrier(const PendingWriteBarrier &write_barrier);
     void ApplyPendingLayoutTransition(const PendingLayoutTransition &layout_transition, ResourceUsageTag layout_transition_tag);
 
-    void ApplySemaphore(const SemaphoreScope &signal, const SemaphoreScope wait);
+    void ApplySemaphore(const SemaphoreScope &signal, const SemaphoreScope &wait);
 
     struct WaitQueueTagPredicate {
         QueueId queue;
